Consolidate E2E tests
